# api.yml
# This is the OpenAPI specification for your service. It is used to generate the client code.
# To generate the server code, use the Gradle task `openApiGenerate`. The
# generated code will be placed in `build/generate-resources/main`.
#
# We will evaluate you based on how well you design your API.
# 1. How well it follows REST principles.
# 2. How easy it is to understand and use.
#
# References
# 1. https://swagger.io/specification/
openapi: "3.0.0"
info:
  version: 1.0.0
  title: User Service
  license:
    name: MIT
servers:
  - url: http://localhost
paths:
  /registration:
    post:
      summary: Sign up to the server
      operationId: registration
      requestBody:
        description: Registration to create user
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Registration"
      responses:
        '201':
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegistrationResponse"
        '400':
          description: Bad Requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /my-profile:
    get:
      summary: Getting detail profile
      operationId: getProfile
      responses:
        '200':
          description: Profile data user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileDataResponse"
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /login:
    post:
      summary: Login to the server
      operationId: login
      requestBody:
        description: Data for login
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Login"
      reponses:
        '201':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        '400':
          description: Login unsuccessful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    Registration:
      type: object
      properties:
        phone_number:
          type: string
        full_name:
          type: string
        password:
          type: string
      required:
        - phone_number
        - full_name
        - password
    Login:
      type: object
      properties:
        phone_number:
          type: string
        password:
          type: string
      required:
        - phone_number
        - password
    RegistrationResponse:
      type: object
      properties:
        id:
          type: integer
          format: int32
        full_name:
          type: string
    ProfileDataResponse:
      type: object
      properties:
        phone_number:
          type: string
        full_name:
          type: string
    ErrorResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string